Change from int16 to int32 for port number assigned to an instance

+++ pkg/controller/dual-pods/controller.go
@@ -303,7 +303,7 @@ type serverData struct {
 	NominalProvidingPodHash string
 
 	// ServerPort is meaningful if NominalProvidingPod is not nil
-	ServerPort int16
+	ServerPort int32
 
 	// UUIDs of the server's GPUs
 	GPUIDs []string
